convert storage service section display to stimulus
