+++ app/views/libraries/_form.html.erb
@@ -5,7 +5,7 @@
     <div class="row mb-3 input-group">
       <%= form.label :storage_service, class: "col-sm-2 col-form-label" %>
       <div class='col-sm-10 ps-0'>
-        <%= form.select :storage_service, Library::STORAGE_SERVICES.map { |it| [translate("libraries.storage_services.%{service}" % {service: it}), it] }, {}, {class: "form-control col-auto form-select", "data-bs-collapse": "collapse"} %>
+        <%= form.select :storage_service, Library::STORAGE_SERVICES.map { |it| [translate("libraries.storage_services.%{service}" % {service: it}), it] }, {}, {class: "form-control col-auto form-select", "data-bs-collapse": "collapse", "data-controller": "storage-service", "data-action": "change->storage-service#onChange"} %>
         <div id="options-filesystem" class="card p-3 collapse show storage-collapse" data-bs-parent="#library_storage_service">
           <div class="row">
             <p><%= t ".filesystem_description_html" %></p>
